+++ tests/Integration/WhereConditions/WhereHasConditionsDirectiveTest.php
@@ -5,6 +5,7 @@ namespace Tests\Integration\WhereConditions;
 use Nuwave\Lighthouse\WhereConditions\WhereConditionsServiceProvider;
 use Tests\DBTestCase;
 use Tests\Utils\Models\Category;
+use Tests\Utils\Models\Location;
 use Tests\Utils\Models\Post;
 use Tests\Utils\Models\Role;
 use Tests\Utils\Models\User;
@@ -42,6 +43,12 @@ class WhereHasConditionsDirectiveTest extends DBTestCase
         name: String!
     }
 
+    type Location {
+        id: ID!
+        parent: Location @belongsTo
+        children: [Location!] @hasMany
+    }
+
     type Query {
         posts(
             hasUser: _ @whereHasConditions(relation: "user")
@@ -65,6 +72,11 @@ class WhereHasConditionsDirectiveTest extends DBTestCase
         withoutRelation(
             hasCompany: _ @whereHasConditions
         ): [User!]! @all
+
+        locations(
+            hasParent: _ @whereHasConditions(columns: ["id"]),
+            hasChildren: _ @whereHasConditions(columns: ["id"])
+        ): [Location!]! @all
     }
     ';
 
@@ -227,6 +239,69 @@ class WhereHasConditionsDirectiveTest extends DBTestCase
 
     public function testWhereHasBelongsToManyOrNestedConditions(): void
     {
+        $parentWithoutPosts = factory(Category::class)->create();
+
+        $categoryWithParentWithoutPosts = factory(Category::class)->create();
+        $categoryWithParentWithoutPosts->parent()->associate($parentWithoutPosts);
+        $categoryWithParentWithoutPosts->save();
+
+        $postWithCategoryWithParentWithoutPosts = factory(Post::class)->create();
+        $postWithCategoryWithParentWithoutPosts->categories()->attach($categoryWithParentWithoutPosts);
+
+        $parentWithPosts = factory(Category::class)->create();
+        $postWithParent = factory(Post::class)->create();
+        $postWithParent->categories()->attach($parentWithoutPosts);
+
+        $categoryWithParentWithPosts = factory(Category::class)->create();
+        $categoryWithParentWithPosts->parent()->associate($parentWithPosts);
+        $categoryWithParentWithPosts->save();
+
+        $postWithCategoryWithParentWithPosts = factory(Post::class)->create();
+        $postWithCategoryWithParentWithPosts->categories()->attach($categoryWithParentWithPosts);
+
+        $parentWithFooPosts = factory(Category::class)->create();
+        /** @var Post $fooPost */
+        $fooPost = factory(Post::class)->make();
+        $fooPost->title = 'foo';
+        $fooPost->save();
+        $fooPost->categories()->attach($parentWithFooPosts);
+
+        $categoryWithParentWithFooPosts = factory(Category::class)->create();
+        $categoryWithParentWithFooPosts->parent()->associate($parentWithFooPosts);
+        $categoryWithParentWithFooPosts->save();
+
+        $postWithCategoryWithParentWithFooPosts = factory(Post::class)->create();
+        $postWithCategoryWithParentWithFooPosts->categories()->attach($categoryWithParentWithFooPosts);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            posts(
+                hasCategories: {
+                    HAS: {
+                        relation: "parent.posts"
+                        condition: {
+                            column: "title"
+                            value: "foo"
+                        }
+                    }
+                }
+            ) {
+                id
+            }
+        }
+        ')->assertExactJson([
+            'data' => [
+                'posts' => [
+                    [
+                        'id' => (string) $postWithCategoryWithParentWithFooPosts->getKey(),
+                    ],
+                ],
+            ],
+        ]);
+    }
+
+    public function testWhereHasNestedRelationWithDotNotation(): void
+    {
         $category1 = factory(Category::class)->create();
 
         $category2 = factory(Category::class)->create();
@@ -309,6 +384,176 @@ class WhereHasConditionsDirectiveTest extends DBTestCase
                     ],
                     [
                         'id' => (string) $post5->getKey(),
+                    ],
+                ],
+            ],
+        ]);
+    }
+
+    public function testWhereConditionsHasNestedTables(): void
+    {
+        $category1 = factory(Category::class)->create();
+
+        $category2 = factory(Category::class)->create();
+        $category2->parent()->associate($category1);
+        $category2->save();
+
+        $category3 = factory(Category::class)->create();
+        $category3->parent()->associate($category2);
+        $category3->save();
+
+        $category4 = factory(Category::class)->create();
+        $category4->parent()->associate($category3);
+        $category4->save();
+
+        $post = factory(Post::class)->create();
+        $post->categories()->attach($category4);
+
+        $this->graphQL(/** @lang GraphQL */ '
+        query ($hasCategories: WhereConditions) {
+            posts(hasCategories: $hasCategories) {
+                id
+                title
+                body
+                categories {
+                    category_id
+                    name
+                    parent {
+                        category_id
+                        name
+                        parent {
+                            category_id
+                            name
+                            parent {
+                                category_id
+                                name
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        ', [
+            'hasCategories' => [
+                'HAS' => [
+                    'relation' => 'parent.parent.parent',
+                    'condition' => [
+                        'column' => 'category_id',
+                        'value' => $category1->category_id,
+                    ],
+                ],
+            ],
+        ])->assertExactJson([
+            'data' => [
+                'posts' => [
+                    [
+                        'id' => (string) $post->id,
+                        'title' => $post->title,
+                        'body' => $post->body,
+                        'categories' => [
+                            [
+                                'category_id' => (string) $category4->category_id,
+                                'name' => $category4->name,
+                                'parent' => [
+                                    'category_id' => (string) $category3->category_id,
+                                    'name' => $category3->name,
+                                    'parent' => [
+                                        'category_id' => (string) $category2->category_id,
+                                        'name' => $category2->name,
+                                        'parent' => [
+                                            'category_id' => (string) $category1->category_id,
+                                            'name' => $category1->name,
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+    }
+
+    public function testWhereHasBelongsToSameTableRelationship(): void
+    {
+        /** @var \Tests\Utils\Models\Location $parent */
+        $parent = factory(Location::class)->create();
+
+        /** @var \Tests\Utils\Models\Location $child */
+        $child = factory(Location::class)->make();
+        $child->parent()->associate($parent);
+        $child->save();
+
+        $this->graphQL(/** @lang GraphQL */ '
+        query ($hasParent: QueryLocationsHasParentWhereHasConditions) {
+            locations(hasParent: $hasParent) {
+                id
+                parent {
+                    id
+                }
+                children {
+                    id
+                }
+            }
+        }
+        ', [
+            'hasParent' => [
+                'column' => 'ID',
+                'value' => $parent->id,
+            ],
+        ])->assertExactJson([
+            'data' => [
+                'locations' => [
+                    [
+                        'id' => (string) $child->id,
+                        'parent' => [
+                            'id' => (string) $parent->id,
+                        ],
+                        'children' => [],
+                    ],
+                ],
+            ],
+        ]);
+    }
+
+    public function testWhereHasHasManySameTableRelationship(): void
+    {
+        /** @var \Tests\Utils\Models\Location $parent */
+        $parent = factory(Location::class)->create();
+
+        /** @var \Tests\Utils\Models\Location $child */
+        $child = factory(Location::class)->make();
+        $child->parent()->associate($parent);
+        $child->save();
+
+        $this->graphQL(/** @lang GraphQL */ '
+        query ($hasChildren: QueryLocationsHasChildrenWhereHasConditions) {
+            locations(hasChildren: $hasChildren) {
+                id
+                parent {
+                    id
+                }
+                children {
+                    id
+                }
+            }
+        }
+        ', [
+            'hasChildren' => [
+                'column' => 'ID',
+                'value' => $child->id,
+            ],
+        ])->assertExactJson([
+            'data' => [
+                'locations' => [
+                    [
+                        'id' => (string) $parent->id,
+                        'parent' => null,
+                        'children' => [
+                            [
+                                'id' => (string) $child->id,
+                            ],
+                        ],
                     ],
                 ],
             ],